2025.09.11.1
IT WORKS! Huge audio manager cleanup, scriptable objects, sounds stop when paused but music keeps playing
