+++ Assets/Audio/AudioSO/CrateHit.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f94cc3d9eb6f3e14d821fee4d0483e9b
+guid: 4671592458c8f3943aa8dbc1bebd37e7
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
