+++ Assets/Audio/AudioSO/Lock.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f94cc3d9eb6f3e14d821fee4d0483e9b
+guid: 042d0b18e18c98148a51f339fbf71fe4
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
