+++ Assets/Audio/AudioSO/Rotation.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f94cc3d9eb6f3e14d821fee4d0483e9b
+guid: 2d22309ee3af60f48862a17c2a6fe177
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
