+++ Assets/Audio/AudioSO/Thrust.asset
@@ -10,11 +10,11 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 0}
-  m_Name: Boost
+  m_Name: Thrust
   m_EditorClassIdentifier: Assembly-CSharp::ContinuousAudio
   mixerGroup: {fileID: 832789883986060108, guid: be51724ba3e3e7f408386242dea02a7f, type: 2}
-  clip: {fileID: 8300000, guid: 622a4fb17c3d77f4c8c70f9cc0bdf667, type: 3}
+  clip: {fileID: 8300000, guid: 9208f5148345c2c47900f4ac0f0ebbaa, type: 3}
   source: {fileID: 0}
   minVolume: 0
   maxVolume: 1
-  restartOnActivation: 0
+  restartOnActivation: 1
