+++ Assets/Audio/AudioSO/Thrust.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f94cc3d9eb6f3e14d821fee4d0483e9b
+guid: 8c26da26ad219ed4ab2b2cc9f4f858e0
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
