+++ Assets/Scripts/Audio/ContinuousAudio.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0a7f6a592ca7da408b09aae03e83035
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
