+++ Assets/Scripts/Audio/OneShotAudio.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Audio;
+
+[CreateAssetMenu(fileName = "OneShotAudioData", menuName = "Audio/OneShot")]
+public class OneShotAudio : AudioData
+{
+    
+}
